+++ CHANGELOG.md
@@ -9,6 +9,9 @@ FaasJS use [Semantic Versioning](https://semver.org/).
 - `faasjs`
   - [Fix] Fix `tsconfig.react.json`.
 
+- `@faasjs/server`
+  - [Feature] Add `middleware` to `Server`.
+
 [`v6.10.0 (2025-03-30)`](https://github.com/faasjs/faasjs/compare/v6.9.0...v6.10.0)
 
 - `@faasjs/server`